Remove separate parse file
With the change it now works on my Windows machine.
Alternative would be to rename the file, but it is small enough to juts
be merged without losing much clarity.
